requirements.txt
aiohttp==3.9.4
APScheduler==3.10.4
motor==3.5.1
pandas
psutil==5.9.8
git+https://github.com/KurimuzonAkuma/pyrogram
pyromod==3.1.6
python-dotenv==1.0.1
TgCrypto==1.2.3
pymongo==4.8.0
async-lru==2.0.4